fix: Remove DialogFragment instead Use AlertDialog
수정사항
1. DialogFragment remove, AlertDialog add
2. AlertDialog Theme custom (버튼 색 kurly_purple로 변경)
3. night theme 삭제 - 항상 light mode로 동작
