--- app/src/main/res/values-night/themes.xml
@@ -1,8 +1,0 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <!-- Base application theme. -->
-    <style name="Theme.ZeroError" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
-        <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">@color/kurly_purple</item>
-        <!-- Customize your theme here. -->
-    </style>
-</resources>
