+++ app/src/main/res/values/themes.xml
@@ -1,8 +1,9 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.ZeroError" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+    <style name="Theme.ZeroError" parent="Theme.MaterialComponents.Light.NoActionBar">
         <!-- Status bar color. -->
         <item name="android:statusBarColor" tools:targetApi="l">@color/kurly_purple</item>
         <!-- Customize your theme here. -->
+        <item name="alertDialogTheme">@style/AlertDialogTheme</item>
     </style>
 </resources>
